Refactored WPBrowser and WPWebDriver (PHP 5.4+ is required)
Moved WPBrowser's methods into WPBrowserMethods trait and use it in both WPBrowser and WPWebDriver
